Remove index.py and req/constructors.py's dependence on download.py.
