+++ src/pip/_internal/operations/prepare.py
@@ -13,12 +13,7 @@ from pip._internal.distributions import (
     make_distribution_for_install_requirement,
 )
 from pip._internal.distributions.installed import InstalledDistribution
-from pip._internal.download import (
-    is_dir_url,
-    is_file_url,
-    unpack_url,
-    url_to_path,
-)
+from pip._internal.download import is_dir_url, is_file_url, unpack_url
 from pip._internal.exceptions import (
     DirectoryUrlHashUnsupported,
     HashUnpinned,
@@ -32,6 +27,7 @@ from pip._internal.utils.logging import indent_log
 from pip._internal.utils.marker_files import write_delete_marker_file
 from pip._internal.utils.misc import display_path, normalize_path
 from pip._internal.utils.typing import MYPY_CHECK_RUNNING
+from pip._internal.utils.urls import url_to_path
 
 if MYPY_CHECK_RUNNING:
     from typing import Optional
