+++ src/pip/_internal/req/constructors.py
@@ -20,15 +20,20 @@ from pip._vendor.packaging.requirements import InvalidRequirement, Requirement
 from pip._vendor.packaging.specifiers import Specifier
 from pip._vendor.pkg_resources import RequirementParseError, parse_requirements
 
-from pip._internal.download import is_archive_file, is_url, url_to_path
 from pip._internal.exceptions import InstallationError
 from pip._internal.models.index import PyPI, TestPyPI
 from pip._internal.models.link import Link
 from pip._internal.pyproject import make_pyproject_path
 from pip._internal.req.req_install import InstallRequirement
-from pip._internal.utils.misc import is_installable_dir, path_to_url
+from pip._internal.utils.misc import (
+    ARCHIVE_EXTENSIONS,
+    is_installable_dir,
+    path_to_url,
+    splitext,
+)
 from pip._internal.utils.typing import MYPY_CHECK_RUNNING
-from pip._internal.vcs import vcs
+from pip._internal.utils.urls import url_to_path
+from pip._internal.vcs import is_url, vcs
 from pip._internal.wheel import Wheel
 
 if MYPY_CHECK_RUNNING:
@@ -45,6 +50,15 @@ __all__ = [
 
 logger = logging.getLogger(__name__)
 operators = Specifier._operators.keys()
+
+
+def is_archive_file(name):
+    # type: (str) -> bool
+    """Return True if `name` is a considered as an archive file."""
+    ext = splitext(name)[1].lower()
+    if ext in ARCHIVE_EXTENSIONS:
+        return True
+    return False
 
 
 def _strip_extras(path):
